+++ docs/developer/FRONTEND_COMPOSABLES.md
@@ -353,7 +353,7 @@ export function useDocumentMetadata(docId: string) {
 For paginated data, use the canonical `usePagination()`:
 
 ```typescript
-import { usePagination } from '@/composables/usePagination'
+import { usePagination } from '@autobot/ui'
 
 export function useDocumentList() {
   const documents = ref<Document[]>([])
@@ -454,11 +454,11 @@ Each composable has a co-located `.examples.md` with usage patterns and edge cas
 - [useConnectionTester.examples.md](../../autobot-frontend/src/composables/useConnectionTester.examples.md)
 - [useErrorHandler.examples.md](../../autobot-frontend/src/composables/useErrorHandler.examples.md)
 - [useErrorHandler.api-migration.md](../../autobot-frontend/src/composables/useErrorHandler.api-migration.md)
-- [useFormValidation.examples.md](../../autobot-frontend/src/composables/useFormValidation.examples.md)
+- [useFormValidation.examples.md](../../libs/autobot-ui/src/composables/useFormValidation.examples.md)
 - [useKeyboard.examples.md](../../autobot-frontend/src/composables/useKeyboard.examples.md)
 - [useLocalStorage.examples.md](../../autobot-frontend/src/composables/useLocalStorage.examples.md)
 - [useModal.examples.md](../../autobot-frontend/src/composables/useModal.examples.md)
-- [usePagination.examples.md](../../autobot-frontend/src/composables/usePagination.examples.md)
+- [usePagination.examples.md](../../libs/autobot-ui/src/composables/usePagination.examples.md)
 - [useTimeout.examples.md](../../autobot-frontend/src/composables/useTimeout.examples.md)
 - [iconMappings.examples.md](../../autobot-frontend/src/utils/iconMappings.examples.md)
